test: skip duration check tests on darwin
These tests will fail on only macos-latest of GitHub Actions.
